implemented custom useForm hook
