+++ client/src/components/CheckoutForm.test.js
@@ -1,9 +1,9 @@
 import React from "react";
-import { render } from "@testing-library/react";
+import { render, getByText, fireEvent, getByTestId} from "@testing-library/react";
 import CheckoutForm from "./CheckoutForm";
 
 // Write up the two tests here and make sure they are testing what the title shows
 
 test("form header renders", () => {});
-
+    const { getByText } = render(<CheckoutForm />)
 test("form shows success message on submit with form details", () => {});
